requirements.txt
pillow-avif-plugin==1.2.0
pillow==8.3.2
numpy==1.19.5
